pyproject.toml
[project]
name = "scalable-textgrad"
version = "0.1.0"
description = "Evolvable agent platform with an Architect service"
authors = [{name = "Codex"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109",
    "uvicorn>=0.24",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "gitpython>=3.1",
    "semver>=3.0",
    "filelock>=3.13",
    "pyyaml>=6.0",
    "opentelemetry-api>=1.22",
    "opentelemetry-sdk>=1.22",
    "python-json-logger>=2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.1.9",
    "mypy>=1.8"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
